@@ -1,27 +1,16 @@
 [build-system]
 build-backend = "setuptools.build_meta"
-requires = [
-  "setuptools",
-]
+requires = [ "setuptools" ]
 
 [project]
 name = "query-exporter"
 description = "Export Prometheus metrics generated from SQL queries"
 readme = "README.rst"
-keywords = [
-  "exporter",
-  "metric",
-  "prometheus",
-  "sql",
-]
-license = {file = "LICENSE.txt"}
-maintainers = [
-    {name = "Alberto Donato", email = "alberto.donato@gmail.com"},
-]
-authors = [
-    {name = "Alberto Donato", email = "alberto.donato@gmail.com"},
-]
-requires-python = ">=3.10"
+keywords = [ "exporter", "metric", "prometheus", "sql" ]
+license = { file = "LICENSE.txt" }
+maintainers = [ { name = "Alberto Donato", email = "alberto.donato@gmail.com" } ]
+authors = [ { name = "Alberto Donato", email = "alberto.donato@gmail.com" } ]
+requires-python = ">=3.11"
 classifiers = [
   "Development Status :: 5 - Production/Stable",
   "Framework :: AsyncIO",
