@@ -1,3 +1,3 @@
 """Export Prometheus metrics generated from SQL queries."""
 
-__version__ = "3.0.0"
+__version__ = "3.2.0"
