@@ -1,43 +1,55 @@
 """Database wrapper."""
 
 import asyncio
+from collections.abc import (
+    Callable,
+    Iterable,
+    Sequence,
+)
+from dataclasses import (
+    InitVar,
+    dataclass,
+    field,
+)
+from functools import partial
 from itertools import chain
 import logging
-import sys
 import re
-from time import perf_counter
-from traceback import format_tb
-from typing import (
-    Any,
-    cast,
-    NamedTuple,
+from threading import (
+    Thread,
+    current_thread,
 )
+from time import (
+    perf_counter,
+    time,
+)
+from types import TracebackType
+import typing as t
 
 from croniter import croniter
+import snowflake.connector
+from snowflake.connector.connection import SnowflakeConnection
+from snowflake.connector.cursor import ResultMetadata
 from sqlalchemy import (
     create_engine,
     event,
     text,
 )
+from sqlalchemy.engine import (
+    Connection,
+    CursorResult,
+    Engine,
+)
 from sqlalchemy.exc import (
     ArgumentError,
     NoSuchModuleError,
 )
-from sqlalchemy_aio import ASYNCIO_STRATEGY
-from sqlalchemy_aio.asyncio import AsyncioEngine
-from sqlalchemy_aio.base import (
-    AsyncConnection,
-    AsyncResultProxy,
-)
-import snowflake.connector
-from snowflake.connector.connection import SnowflakeConnection
+from sqlalchemy.pool import NullPool
+from sqlalchemy.sql.elements import TextClause
+import structlog
 
-#: Timeout for a query
+# Timeout for a query
 QueryTimeout = int | float
-
-
-#: Label used to tag metrics by database
-DATABASE_LABEL = "database"
 
 
 class DataBaseError(Exception):
@@ -46,7 +58,7 @@ class DataBaseError(Exception):
     if `fatal` is True, it means the Query will never succeed.
     """
 
-    def __init__(self, message: str, fatal: bool = False):
+    def __init__(self, message: str, fatal: bool = False) -> None:
         super().__init__(message)
         self.fatal = fatal
 
@@ -62,16 +74,11 @@ class DataBaseQueryError(DataBaseError):
 class QueryTimeoutExpired(Exception):
     """Query execution timeout expired."""
 
-    def __init__(self, query_name: str, timeout: QueryTimeout):
-        super().__init__(
-            f'Execution for query "{query_name}" expired after {timeout} seconds'
-        )
-
 
 class InvalidResultCount(Exception):
     """Number of results from a query don't match metrics count."""
 
-    def __init__(self, expected: int, got: int):
+    def __init__(self, expected: int, got: int) -> None:
         super().__init__(
             f"Wrong result count from query: expected {expected}, got {got}"
         )
@@ -94,7 +101,7 @@ class InvalidResultColumnNames(Exception):
 class InvalidQueryParameters(Exception):
     """Query parameter names don't match those in query SQL."""
 
-    def __init__(self, query_name: str):
+    def __init__(self, query_name: str) -> None:
         super().__init__(
             f'Parameters for query "{query_name}" don\'t match those from SQL'
         )
@@ -103,13 +110,13 @@ class InvalidQueryParameters(Exception):
 class InvalidQuerySchedule(Exception):
     """Query schedule is wrong or both schedule and interval specified."""
 
-    def __init__(self, query_name: str, message: str):
+    def __init__(self, query_name: str, message: str) -> None:
         super().__init__(
             f'Invalid schedule for query "{query_name}": {message}'
         )
 
 
-# database errors that mean the query won't ever succeed.  Not all possible
+# Database errors that mean the query won't ever succeed.  Not all possible
 # fatal errors are tracked here, because some DBAPI errors can happen in
 # circumstances which can be fatal or not.  Since there doesn't seem to be a
 # reliable way to know, there might be cases when a query will never succeed
@@ -117,10 +124,37 @@ class InvalidQuerySchedule(Exception):
 FATAL_ERRORS = (InvalidResultCount, InvalidResultColumnNames)
 
 
-def create_db_engine(dsn: str, **kwargs) -> AsyncioEngine:
-    """Create the database engine, validating the DSN"""
+class DatabaseConfigError(Exception):
+    """Database Configuration is invalid."""
+
+
+@dataclass(frozen=True)
+class DataBaseConfig:
+    """Configuration for a database."""
+
+    name: str
+    dsn: str
+    conn_type: str = "generic"
+    connect_sql: list[str] = field(default_factory=list)
+    labels: dict[str, str] = field(default_factory=dict)
+    keep_connected: bool = True
+    autocommit: bool = True
+
+    def __post_init__(self) -> None:
+        # raise DatabaseError error if the DSN in invalid
+        try:
+            if self.conn_type == "snowflake":
+                create_snowflake_connection(self.dsn)
+            else:
+                create_db_engine(self.dsn)
+        except DataBaseError as e:
+            raise DatabaseConfigError(str(e))
+
+
+def create_db_engine(dsn: str) -> Engine:
+    """Create the database engine, validating the DSN."""
     try:
-        return create_engine(dsn, **kwargs)
+        return create_engine(dsn, poolclass=NullPool)
     except ImportError as error:
         raise DataBaseError(f'module "{error.name}" not found')
     except (ArgumentError, ValueError, NoSuchModuleError):
@@ -128,8 +162,10 @@ def create_db_engine(dsn: str, **kwargs) -> AsyncioEngine:
 
 
 def create_snowflake_connection(dsn: str) -> SnowflakeConnection:
-    connection_config = re.search('(snowflake:\/\/)(.+)(:)(.+)(@)(.+)', dsn)
+    connection_config = re.search(r"(snowflake:\/\/)(.+)(:)(.+)(@)(.+)", dsn)
 
+    if not connection_config:
+        raise ValueError("not a valid snowflake connection string")
     return snowflake.connector.connect(
         user=connection_config.group(2),
         password=connection_config.group(4),
@@ -137,78 +173,86 @@ def create_snowflake_connection(dsn: str) -> SnowflakeConnection:
     )
 
 
-class QueryMetric(NamedTuple):
+class QueryMetric(t.NamedTuple):
     """Metric details for a Query."""
 
     name: str
-    labels: list[str]
+    labels: Iterable[str]
 
 
-class QueryResults(NamedTuple):
+class QueryResults(t.NamedTuple):
     """Results of a database query."""
 
     keys: list[str]
-    rows: list[tuple]
+    rows: Sequence[Sequence[t.Any]]
+    timestamp: float | None = None
     latency: float | None = None
 
     @classmethod
-    async def from_results(cls, results: AsyncResultProxy):
+    def from_result(cls, result: CursorResult[t.Any]) -> t.Self:
         """Return a QueryResults from results for a query."""
-        conn_info = results._result_proxy.connection.info
-        latency = conn_info.get("query_latency", None)
-        return cls(
-            await results.keys(), await results.fetchall(), latency=latency
-        )
+        timestamp = time()
+        keys: list[str] = []
+        rows: Sequence[Sequence[t.Any]] = []
+        if result.returns_rows:
+            keys, rows = list(result.keys()), result.all()
+        latency = result.connection.info.get("query_latency", None)
+        return cls(keys, rows, timestamp=timestamp, latency=latency)
 
     @classmethod
-    def from_snowflake(cls, keys, results):
+    def from_snowflake(
+        cls,
+        keys: list[ResultMetadata],
+        results: list[tuple[t.Any, ...] | dict[t.Any, t.Any]],
+    ) -> "QueryResults":
         return cls(
             [key.name.lower() for key in keys],
-            [result for result in results]
+            [result for result in results],  # type: ignore[misc]
         )
 
 
-class MetricResult(NamedTuple):
+class MetricResult(t.NamedTuple):
     """A result for a metric from a query."""
 
     metric: str
-    value: Any
+    value: t.Any
     labels: dict[str, str]
 
 
-class MetricResults(NamedTuple):
+class MetricResults(t.NamedTuple):
     """Collection of metric results for a query."""
 
     results: list[MetricResult]
+    timestamp: float | None = None
     latency: float | None = None
 
 
+@dataclass
 class Query:
     """Query definition and configuration."""
 
-    def __init__(
-        self,
-        name: str,
-        databases: list[str],
-        metrics: list[QueryMetric],
-        sql: str,
-        parameters: dict[str, Any] | None = None,
-        timeout: QueryTimeout | None = None,
-        interval: int | None = None,
-        schedule: str | None = None,
-        config_name: str | None = None,
-    ):
-        self.name = name
-        self.databases = databases
-        self.metrics = metrics
-        self.sql = sql
-        self.parameters = parameters or {}
-        self.timeout = timeout
-        self.interval = interval
-        self.schedule = schedule
-        self.config_name = config_name or name
+    name: str
+    databases: list[str]
+    metrics: list[QueryMetric]
+    sql: str
+    timeout: QueryTimeout | None = None
+    interval: int | None = None
+    schedule: str | None = None
+
+    parameter_sets: InitVar[list[dict[str, t.Any]] | None] = None
+    executions: list["QueryExecution"] = field(init=False, compare=False)
+
+    def __post_init__(
+        self, parameter_sets: list[dict[str, t.Any]] | None
+    ) -> None:
         self._check_schedule()
-        self._check_query_parameters()
+        if not parameter_sets:
+            self.executions = [QueryExecution(self.name, self)]
+        else:
+            self.executions = [
+                QueryExecution(f"{self.name}[params{index}]", self, parameters)
+                for index, parameters in enumerate(parameter_sets, 1)
+            ]
 
     @property
     def timed(self) -> bool:
@@ -231,7 +275,7 @@ class Query:
         if len(expected_keys) != len(result_keys):
             raise InvalidResultCount(len(expected_keys), len(result_keys))
         if result_keys != expected_keys:
-            raise InvalidResultColumnNames(result_keys, expected_keys)
+            raise InvalidResultColumnNames(expected_keys, result_keys)
         results = []
         for row in query_results.rows:
             values = dict(zip(query_results.keys, row))
@@ -243,9 +287,13 @@ class Query:
                 )
                 results.append(metric_result)
 
-        return MetricResults(results, latency=query_results.latency)
+        return MetricResults(
+            results,
+            timestamp=query_results.timestamp,
+            latency=query_results.latency,
+        )
 
-    def _check_schedule(self):
+    def _check_schedule(self) -> None:
         if self.interval and self.schedule:
             raise InvalidQuerySchedule(
                 self.name, "both interval and schedule specified"
@@ -253,11 +301,57 @@ class Query:
         if self.schedule and not croniter.is_valid(self.schedule):
             raise InvalidQuerySchedule(self.name, "invalid schedule format")
 
-    def _check_query_parameters(self):
-        expr = text(self.sql)
+
+@dataclass(frozen=True)
+class QueryExecution:
+    """A single execution configuration for a query, with parameters."""
+
+    name: str
+    query: Query
+    parameters: dict[str, t.Any] = field(default_factory=dict)
+
+    def __post_init__(self) -> None:
+        self._check_query_parameters()
+
+    def _check_query_parameters(self) -> None:
+        expr = text(self.query.sql)
         query_params = set(expr.compile().params)
         if set(self.parameters) != query_params:
             raise InvalidQueryParameters(self.name)
+
+
+class WorkerAction:
+    """An action to be called in the worker thread."""
+
+    def __init__(
+        self, func: Callable[..., t.Any], *args: t.Any, **kwargs: t.Any
+    ) -> None:
+        self._func = partial(func, *args, **kwargs)
+        self._loop = asyncio.get_event_loop()
+        self._future = self._loop.create_future()
+
+    def __str__(self) -> str:
+        return self._func.func.__name__
+
+    __repr__ = __str__
+
+    def __call__(self) -> None:
+        """Call the action asynchronously in a thread-safe way."""
+        try:
+            result = self._func()
+        except Exception as e:
+            self._call_threadsafe(self._future.set_exception, e)
+        else:
+            self._call_threadsafe(self._future.set_result, result)
+
+    async def result(self) -> t.Any:
+        """Wait for completion and return the action result."""
+        return await self._future
+
+    def _call_threadsafe(
+        self, call: Callable[..., t.Any], *args: t.Any
+    ) -> None:
+        self._loop.call_soon_threadsafe(partial(call, *args))
 
 
 class SnowflakeDataBase:
@@ -265,23 +359,30 @@ class SnowflakeDataBase:
 
     def __init__(
         self,
-        config,
+        config: DataBaseConfig,
         logger: logging.Logger = logging.getLogger(),
-    ):
+    ) -> None:
         self.config = config
         self.logger = logger
-        self._engine: SnowflakeConnection = create_snowflake_connection(self.config.dsn)
+        self._engine: SnowflakeConnection = create_snowflake_connection(
+            self.config.dsn
+        )
 
-    async def __aenter__(self):
+    async def __aenter__(self) -> "SnowflakeDataBase":
         if not self._engine:
-            self._engine: SnowflakeConnection = create_snowflake_connection(self.config.dsn)
+            self._engine: SnowflakeConnection = create_snowflake_connection(  # type: ignore[no-redef]
+                self.config.dsn
+            )
         return self
 
-    async def __aexit__(self, exc_type, exc_value, traceback):
+    async def __aexit__(
+        self, exc_type: t.Any, exc_value: t.Any, traceback: t.Any
+    ) -> None:
         await self.close()
 
-    async def close(self):
-        await self._engine.close()
+    async def close(self) -> None:
+        self._engine.close()
+        return
 
     async def execute(self, query: Query) -> MetricResults:
         """Execute a query."""
@@ -292,9 +393,11 @@ class SnowflakeDataBase:
         try:
             cur = self._engine.cursor()
             cur.execute_async(query.sql)
-            cur.get_results_from_sfqid(cur.sfqid)
+            cur.get_results_from_sfqid(cur.sfqid)  # type: ignore[arg-type]
             results = cur.fetchall()
-            return query.results(QueryResults.from_snowflake(cur.description, results))
+            return query.results(
+                QueryResults.from_snowflake(cur.description, results)  # type: ignore[arg-type]
+            )
         except Exception as error:
             message = str(error).strip()
             self.logger.error(
@@ -305,91 +408,218 @@ class SnowflakeDataBase:
             assert self._pending_queries >= 0, "pending queries is negative"
             self._pending_queries -= 1
             if not self.config.keep_connected and not self._pending_queries:
-                self.close()
+                await self.close()
+
+
+class DataBaseConnection:
+    """A database to perform Queries."""
+
+    _conn: Connection | None = None
+    _worker: Thread | None = None
+
+    def __init__(
+        self,
+        dbname: str,
+        engine: Engine,
+        logger: structlog.stdlib.BoundLogger | None = None,
+    ) -> None:
+        self.dbname = dbname
+        self.engine = engine
+        self.logger = logger or structlog.get_logger()
+        self._loop = asyncio.get_event_loop()
+        self._queue: asyncio.Queue[WorkerAction] = asyncio.Queue()
+
+    @property
+    def connected(self) -> bool:
+        """Whether the connection is open."""
+        return self._conn is not None
+
+    async def open(self) -> None:
+        """Open the connection."""
+        if self.connected:
+            return
+
+        self._create_worker()
+        await self._call_in_thread(self._connect)
+
+    async def close(self) -> None:
+        """Close the connection."""
+        if not self.connected:
+            return
+
+        await self._call_in_thread(self._close)
+        self._terminate_worker()
+
+    async def execute(
+        self,
+        statement: TextClause,
+        parameters: dict[str, t.Any] | None = None,
+    ) -> QueryResults:
+        """Execute a query, returning results."""
+        if parameters is None:
+            parameters = {}
+        return t.cast(
+            QueryResults,
+            await self._call_in_thread(self._execute, statement, parameters),
+        )
+
+    async def execute_many(
+        self,
+        statements: list[TextClause],
+    ) -> None:
+        """Execute multiple statements."""
+        await self._call_in_thread(self._execute_many, statements)
+
+    def _create_worker(self) -> None:
+        assert not self._worker
+        self._worker = Thread(
+            target=self._run, name=f"DataBase-{self.dbname}", daemon=True
+        )
+        self._worker.start()
+
+    def _terminate_worker(self) -> None:
+        assert self._worker
+        self._worker.join()
+        self._worker = None
+
+    def _connect(self) -> None:
+        self._conn = self.engine.connect()
+
+    def _execute(
+        self, statement: TextClause, parameters: dict[str, t.Any]
+    ) -> QueryResults:
+        assert self._conn
+        with self._conn.begin():
+            result = self._conn.execute(statement, parameters)
+            return QueryResults.from_result(result)
+
+    def _execute_many(self, statements: list[TextClause]) -> None:
+        assert self._conn
+        with self._conn.begin():
+            for statement in statements:
+                self._conn.execute(statement)
+
+    def _close(self) -> None:
+        assert self._conn
+        self._conn.detach()
+        self._conn.close()
+        self._conn = None
+
+    def _run(self) -> None:
+        """The worker thread function."""
+        logger = self.logger.bind(worker_id=current_thread().native_id)
+        logger.debug("start")
+        while True:
+            future = asyncio.run_coroutine_threadsafe(
+                self._queue.get(), self._loop
+            )
+            action = future.result()
+            logger.debug("action received", action=str(action))
+            action()
+            self._loop.call_soon_threadsafe(self._queue.task_done)
+            if self._conn is None:
+                # the connection has been closed, exit the thread
+                logger.debug("shutdown")
+                return
+
+    async def _call_in_thread(
+        self, func: Callable[..., t.Any], *args: t.Any, **kwargs: t.Any
+    ) -> t.Any:
+        """Call a sync action in the worker thread."""
+        call = WorkerAction(func, *args, **kwargs)
+        await self._queue.put(call)
+        return await call.result()
 
 
 class DataBase:
     """A database to perform Queries."""
 
-    _engine: AsyncioEngine
-    _conn: AsyncConnection | None = None
+    _conn: DataBaseConnection
     _pending_queries: int = 0
 
     def __init__(
         self,
-        config,
-        logger: logging.Logger = logging.getLogger(),
-    ):
+        config: DataBaseConfig,
+        logger: structlog.stdlib.BoundLogger | None = None,
+    ) -> None:
         self.config = config
-        self.logger = logger
+        if logger is None:
+            logger = structlog.get_logger()
+        self.logger = logger.bind(database=self.config.name)
         self._connect_lock = asyncio.Lock()
-        self._engine = create_db_engine(
-            self.config.dsn,
-            strategy=ASYNCIO_STRATEGY,
-            execution_options={"autocommit": self.config.autocommit},
-        )
+        engine = create_db_engine(self.config.dsn)
+        self._conn = DataBaseConnection(self.config.name, engine, self.logger)
+        self._setup_query_latency_tracking(engine)
 
-        self._setup_query_latency_tracking()
-
-    async def __aenter__(self):
+    async def __aenter__(self) -> t.Self:
         await self.connect()
         return self
 
-    async def __aexit__(self, exc_type, exc_value, traceback):
+    async def __aexit__(
+        self,
+        exc_type: type,
+        exc_value: Exception,
+        traceback: TracebackType,
+    ) -> None:
         await self.close()
 
     @property
     def connected(self) -> bool:
         """Whether the database is connected."""
-        return self._conn is not None
+        return self._conn.connected
 
-    async def connect(self):
+    async def connect(self) -> None:
         """Connect to the database."""
         async with self._connect_lock:
             if self.connected:
                 return
 
             try:
-                self._conn = await self._engine.connect()
+                await self._conn.open()
             except Exception as error:
                 raise self._db_error(error, exc_class=DataBaseConnectError)
 
-            self.logger.debug(f'connected to database "{self.config.name}"')
-            for sql in self.config.connect_sql:
+            self.logger.debug("connected")
+            if self.config.connect_sql:
                 try:
-                    await self.execute_sql(sql)
+                    await self._conn.execute_many(
+                        [text(sql) for sql in self.config.connect_sql]
+                    )
                 except Exception as error:
                     await self._close()
                     raise self._db_error(
-                        f'failed executing query "{sql}": {error}',
+                        f"failed executing connect SQL: {error}",
                         exc_class=DataBaseQueryError,
                     )
 
-    async def close(self):
+    async def close(self) -> None:
         """Close the database connection."""
         async with self._connect_lock:
             if not self.connected:
                 return
             await self._close()
 
-    async def execute(self, query: Query) -> MetricResults:
+    async def execute(self, query_execution: QueryExecution) -> MetricResults:
         """Execute a query."""
         await self.connect()
-        self.logger.debug(
-            f'running query "{query.name}" on database "{self.config.name}"'
-        )
+        self.logger.debug("run query", query=query_execution.name)
         self._pending_queries += 1
-        self._conn: AsyncConnection
+        query = query_execution.query
         try:
-            result = await self._execute_query(query)
-            return query.results(await QueryResults.from_results(result))
-        except asyncio.TimeoutError:
-            raise self._query_timeout_error(
-                query.name, cast(QueryTimeout, query.timeout)
+            query_results = await self.execute_sql(
+                query.sql,
+                parameters=query_execution.parameters,
+                timeout=query.timeout,
             )
+            return query.results(query_results)
+        except TimeoutError:
+            self.logger.warning("query timeout", query=query_execution.name)
+            raise QueryTimeoutExpired()
         except Exception as error:
             raise self._query_db_error(
-                query.name, error, fatal=isinstance(error, FATAL_ERRORS)
+                query_execution.name,
+                error,
+                fatal=isinstance(error, FATAL_ERRORS),
             )
         finally:
             assert self._pending_queries >= 0, "pending queries is negative"
@@ -400,45 +630,31 @@ class DataBase:
     async def execute_sql(
         self,
         sql: str,
-        parameters: dict[str, Any] | None = None,
+        parameters: dict[str, t.Any] | None = None,
         timeout: QueryTimeout | None = None,
-    ) -> AsyncResultProxy:
+    ) -> QueryResults:
         """Execute a raw SQL query."""
-        if parameters is None:
-            parameters = {}
-        self._conn: AsyncConnection
         return await asyncio.wait_for(
             self._conn.execute(text(sql), parameters),
             timeout=timeout,
         )
 
-    async def _execute_query(self, query: Query) -> AsyncResultProxy:
-        """Execute a query."""
-        return await self.execute_sql(
-            query.sql, parameters=query.parameters, timeout=query.timeout
-        )
-
-    async def _close(self):
+    async def _close(self) -> None:
         # ensure the connection with the DB is actually closed
-        self._conn.sync_connection.detach()
         await self._conn.close()
-        self._conn = None
-        self._pending_queries = 0
-        self.logger.debug(f'disconnected from database "{self.config.name}"')
+        self.logger.debug("disconnected")
 
-    def _setup_query_latency_tracking(self):
-        engine = self._engine.sync_engine
-
-        @event.listens_for(engine, "before_cursor_execute")
+    def _setup_query_latency_tracking(self, engine: Engine) -> None:
+        @event.listens_for(engine, "before_cursor_execute")  # type: ignore
         def before_cursor_execute(
             conn, cursor, statement, parameters, context, executemany
-        ):
+        ) -> None:
             conn.info["query_start_time"] = perf_counter()
 
-        @event.listens_for(engine, "after_cursor_execute")
+        @event.listens_for(engine, "after_cursor_execute")  # type: ignore
         def after_cursor_execute(
             conn, cursor, statement, parameters, context, executemany
-        ):
+        ) -> None:
             conn.info["query_latency"] = perf_counter() - conn.info.pop(
                 "query_start_time"
             )
@@ -451,20 +667,10 @@ class DataBase:
     ) -> DataBaseError:
         """Create and log a DataBaseError for a failed query."""
         message = self._error_message(error)
-        self.logger.error(
-            f'query "{query_name}" on database "{self.config.name}" failed: '
-            + message
+        self.logger.exception(
+            "query failed", query=query_name, error=message, exception=error
         )
-        _, _, traceback = sys.exc_info()
-        self.logger.debug("".join(format_tb(traceback)))
         return DataBaseQueryError(message, fatal=fatal)
-
-    def _query_timeout_error(
-        self, query_name: str, timeout: QueryTimeout
-    ) -> QueryTimeoutExpired:
-        error = QueryTimeoutExpired(query_name, timeout)
-        self.logger.warning(str(error))
-        raise error
 
     def _db_error(
         self,
@@ -474,9 +680,7 @@ class DataBase:
     ) -> DataBaseError:
         """Create and log a DataBaseError."""
         message = self._error_message(error)
-        self.logger.error(
-            f'error from database "{self.config.name}": {message}'
-        )
+        self.logger.exception("database error", exception=error)
         return exc_class(message, fatal=fatal)
 
     def _error_message(self, error: str | Exception) -> str:
