@@ -1,38 +1,28 @@
 import asyncio
-from collections import defaultdict
-from datetime import datetime
+from collections.abc import AsyncIterator, Callable, Iterator
 from decimal import Decimal
-import logging
 from pathlib import Path
-import re
+import typing as t
+from unittest.mock import ANY
 
 from prometheus_aioexporter import MetricsRegistry
 import pytest
+from pytest_mock import MockerFixture
+from pytest_structlog import StructuredLogCapture
+from sqlalchemy.sql.elements import TextClause
 import yaml
 
-from query_exporter import loop
-from query_exporter.config import (
-    DataBaseConfig,
-    load_config,
-)
-from query_exporter.db import DataBase
+from query_exporter.config import load_config
+from query_exporter.db import DataBase, DataBaseConfig
+from query_exporter.loop import MetricsLastSeen, QueryLoop
 
+from .conftest import QueryTracker, metric_values
 
-class re_match:
-    """Assert that comparison matches the specified regexp."""
-
-    def __init__(self, pattern, flags=0):
-        self._re = re.compile(pattern, flags)
-
-    def __eq__(self, string):
-        return bool(self._re.match(string))
-
-    def __repr__(self):
-        return self._re.pattern  # pragma: nocover
+AdvanceTime = Callable[[float], t.Awaitable[None]]
 
 
 @pytest.fixture
-def config_data():
+def config_data() -> Iterator[dict[str, t.Any]]:
     yield {
         "databases": {"db": {"dsn": "sqlite://"}},
         "metrics": {"m": {"type": "gauge"}},
@@ -48,21 +38,25 @@ def config_data():
 
 
 @pytest.fixture
-def registry():
+def registry() -> Iterator[MetricsRegistry]:
     yield MetricsRegistry()
 
 
+MakeQueryLoop = Callable[[], QueryLoop]
+
+
 @pytest.fixture
-async def make_query_loop(tmp_path, config_data, registry):
+async def make_query_loop(
+    tmp_path: Path, config_data: dict[str, t.Any], registry: MetricsRegistry
+) -> AsyncIterator[MakeQueryLoop]:
     query_loops = []
 
-    def make_loop():
+    def make_loop() -> QueryLoop:
         config_file = tmp_path / "config.yaml"
         config_file.write_text(yaml.dump(config_data), "utf-8")
-        with config_file.open() as fh:
-            config = load_config(fh, logging.getLogger())
+        config = load_config([config_file])
         registry.create_metrics(config.metrics.values())
-        query_loop = loop.QueryLoop(config, registry, logging)
+        query_loop = QueryLoop(config, registry)
         query_loops.append(query_loop)
         return query_loop
 
@@ -74,39 +68,13 @@ async def make_query_loop(tmp_path, config_data, registry):
 
 
 @pytest.fixture
-async def query_loop(make_query_loop):
+async def query_loop(
+    make_query_loop: MakeQueryLoop,
+) -> AsyncIterator[QueryLoop]:
     yield make_query_loop()
 
 
-@pytest.fixture
-def metrics_expiration():
-    yield {
-        "m1": 50,
-        "m2": 100,
-        "m3": None,
-    }
-
-
-def metric_values(metric, by_labels=()):
-    """Return values for the metric."""
-    if metric._type == "gauge":
-        suffix = ""
-    elif metric._type == "counter":
-        suffix = "_total"
-
-    values = defaultdict(list)
-    for sample_suffix, labels, value, *_ in metric._samples():
-        if sample_suffix == suffix:
-            if by_labels:
-                label_values = tuple(labels[label] for label in by_labels)
-                values[label_values] = value
-            else:
-                values[sample_suffix].append(value)
-
-    return values if by_labels else values[suffix]
-
-
-async def run_queries(db_file: Path, *queries: str):
+async def run_queries(db_file: Path, *queries: str) -> None:
     config = DataBaseConfig(name="db", dsn=f"sqlite:///{db_file}")
     async with DataBase(config) as db:
         for query in queries:
@@ -114,9 +82,8 @@ async def run_queries(db_file: Path, *queries: str):
 
 
 class TestMetricsLastSeen:
-    def test_update(self, metrics_expiration):
-        """Last seen times are tracked for each series of metrics with expiration."""
-        last_seen = loop.MetricsLastSeen(metrics_expiration)
+    def test_update(self) -> None:
+        last_seen = MetricsLastSeen({"m1": 50, "m2": 100})
         last_seen.update("m1", {"l1": "v1", "l2": "v2"}, 100)
         last_seen.update("m1", {"l1": "v3", "l2": "v4"}, 200)
         last_seen.update("other", {"l3": "v100"}, 300)
@@ -127,43 +94,65 @@ class TestMetricsLastSeen:
             }
         }
 
-    def test_update_label_values_sorted_by_name(self, metrics_expiration):
-        """Last values are sorted by label names."""
-        last_seen = loop.MetricsLastSeen(metrics_expiration)
+    def test_update_label_values_sorted_by_name(self) -> None:
+        last_seen = MetricsLastSeen({"m1": 50})
         last_seen.update("m1", {"l2": "v2", "l1": "v1"}, 100)
         assert last_seen._last_seen == {"m1": {("v1", "v2"): 100}}
 
-    def test_expire_series(self, metrics_expiration):
-        """Expired metric series are returned and removed."""
-        last_seen = loop.MetricsLastSeen(metrics_expiration)
+    def test_expire_series_not_expired(self) -> None:
+        last_seen = MetricsLastSeen({"m1": 50})
+        last_seen.update("m1", {"l1": "v1", "l2": "v2"}, 10)
+        last_seen.update("m1", {"l1": "v3", "l2": "v4"}, 20)
+        assert last_seen.expire_series(30) == {}
+        assert last_seen._last_seen == {
+            "m1": {
+                ("v1", "v2"): 10,
+                ("v3", "v4"): 20,
+            }
+        }
+
+    def test_expire_series(self) -> None:
+        last_seen = MetricsLastSeen({"m1": 50, "m2": 100})
         last_seen.update("m1", {"l1": "v1", "l2": "v2"}, 10)
         last_seen.update("m1", {"l1": "v3", "l2": "v4"}, 100)
         last_seen.update("m2", {"l3": "v100"}, 100)
-        expired = last_seen.expire_series(120)
-        assert expired == {"m1": [("v1", "v2")], "m2": []}
+        assert last_seen.expire_series(120) == {"m1": [("v1", "v2")]}
         assert last_seen._last_seen == {
             "m1": {("v3", "v4"): 100},
             "m2": {("v100",): 100},
         }
 
+    def test_expire_no_labels(self) -> None:
+        last_seen = MetricsLastSeen({"m1": 50})
+        last_seen.update("m1", {}, 10)
+        expired = last_seen.expire_series(120)
+        assert expired == {"m1": [()]}
+        assert last_seen._last_seen == {}
 
-@pytest.mark.asyncio
+
 class TestQueryLoop:
-    async def test_start(self, query_loop):
-        """The start method starts timed calls for queries."""
+    async def test_start(
+        self,
+        query_tracker: QueryTracker,
+        query_loop: QueryLoop,
+    ) -> None:
         await query_loop.start()
         timed_call = query_loop._timed_calls["q"]
         assert timed_call.running
+        await query_tracker.wait_results()
 
-    async def test_stop(self, query_loop):
-        """The stop method stops timed calls for queries."""
+    async def test_stop(self, query_loop: QueryLoop) -> None:
         await query_loop.start()
         timed_call = query_loop._timed_calls["q"]
         await query_loop.stop()
         assert not timed_call.running
 
-    async def test_run_query(self, query_tracker, query_loop, registry):
-        """Queries are run and update metrics."""
+    async def test_run_query(
+        self,
+        query_tracker: QueryTracker,
+        query_loop: QueryLoop,
+        registry: MetricsRegistry,
+    ) -> None:
         await query_loop.start()
         await query_tracker.wait_results()
         # the metric is updated
@@ -177,25 +166,27 @@ class TestQueryLoop:
 
     async def test_run_scheduled_query(
         self,
-        mocker,
-        event_loop,
-        advance_time,
-        query_tracker,
-        registry,
-        config_data,
-        make_query_loop,
-    ):
-        """Queries are run and update metrics."""
+        mocker: MockerFixture,
+        advance_time: AdvanceTime,
+        query_tracker: QueryTracker,
+        registry: MetricsRegistry,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+    ) -> None:
+        event_loop = asyncio.get_running_loop()
 
-        def croniter(*args):
+        def croniter(*args: t.Any) -> Iterator[float]:
             while True:
                 # sync croniter time with the loop one
                 yield event_loop.time() + 60
 
-        mock_croniter = mocker.patch.object(loop, "croniter")
+        mock_croniter = mocker.patch("query_exporter.loop.croniter")
         mock_croniter.side_effect = croniter
         # ensure that both clocks advance in sync
-        mocker.patch.object(loop.time, "time", lambda: event_loop.time())
+        mocker.patch(
+            "query_exporter.loop.time.time",
+            lambda: event_loop.time(),
+        )
 
         del config_data["queries"]["q"]["interval"]
         config_data["queries"]["q"]["schedule"] = "*/2 * * * *"
@@ -204,9 +195,12 @@ class TestQueryLoop:
         mock_croniter.assert_called_once()
 
     async def test_run_query_with_parameters(
-        self, query_tracker, registry, config_data, make_query_loop
-    ):
-        """Queries are run with declared parameters."""
+        self,
+        query_tracker: QueryTracker,
+        registry: MetricsRegistry,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+    ) -> None:
         config_data["metrics"]["m"]["type"] = "counter"
         config_data["metrics"]["m"]["labels"] = ["l"]
         config_data["queries"]["q"]["sql"] = "SELECT :param AS m, :label as l"
@@ -230,9 +224,12 @@ class TestQueryLoop:
         }
 
     async def test_run_query_null_value(
-        self, query_tracker, registry, config_data, make_query_loop
-    ):
-        """A null value in query results is treated like a zero."""
+        self,
+        query_tracker: QueryTracker,
+        registry: MetricsRegistry,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+    ) -> None:
         config_data["queries"]["q"]["sql"] = "SELECT NULL AS m"
         query_loop = make_query_loop()
         await query_loop.start()
@@ -240,13 +237,25 @@ class TestQueryLoop:
         metric = registry.get_metric("m")
         assert metric_values(metric) == [0]
 
-    async def test_run_query_counter_no_increment(
-        self, query_tracker, registry, config_data, make_query_loop
-    ):
-        """If increment is set to False, counter is set to the new value."""
+    @pytest.mark.parametrize(
+        "increment,value",
+        [
+            (True, 30.0),
+            (False, 20.0),
+        ],
+    )
+    async def test_run_query_counter(
+        self,
+        query_tracker: QueryTracker,
+        registry: MetricsRegistry,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+        increment: bool,
+        value: float,
+    ) -> None:
         config_data["metrics"]["m"]["type"] = "counter"
-        config_data["metrics"]["m"]["increment"] = False
         config_data["queries"]["q"]["sql"] = "SELECT :param AS m"
+        config_data["metrics"]["m"]["increment"] = increment
         config_data["queries"]["q"]["parameters"] = [
             {"param": 10.0},
             {"param": 20.0},
@@ -256,12 +265,15 @@ class TestQueryLoop:
         await query_tracker.wait_results()
         # the metric is updated
         metric = registry.get_metric("m")
-        assert metric_values(metric) == [20.0]
+        assert metric_values(metric) == [value]
 
     async def test_run_query_metrics_with_database_labels(
-        self, query_tracker, registry, config_data, make_query_loop
-    ):
-        """If databases have extra labels, they're set for metrics."""
+        self,
+        query_tracker: QueryTracker,
+        registry: MetricsRegistry,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+    ) -> None:
         config_data["databases"] = {
             "db1": {"dsn": "sqlite://", "labels": {"l1": "v1", "l2": "v2"}},
             "db2": {"dsn": "sqlite://", "labels": {"l1": "v3", "l2": "v4"}},
@@ -277,9 +289,10 @@ class TestQueryLoop:
         }
 
     async def test_update_metric_decimal_value(
-        self, registry, make_query_loop
-    ):
-        """A Decimal value in query results is converted to float."""
+        self,
+        registry: MetricsRegistry,
+        make_query_loop: MakeQueryLoop,
+    ) -> None:
         db = DataBase(DataBaseConfig(name="db", dsn="sqlite://"))
         query_loop = make_query_loop()
         query_loop._update_metric(db, "m", Decimal("100.123"))
@@ -288,48 +301,71 @@ class TestQueryLoop:
         assert value == 100.123
         assert isinstance(value, float)
 
-    async def test_run_query_log(self, caplog, query_tracker, query_loop):
-        """Debug messages are logged on query execution."""
-        caplog.set_level(logging.DEBUG)
+    async def test_run_query_log(
+        self,
+        log: StructuredLogCapture,
+        query_tracker: QueryTracker,
+        query_loop: QueryLoop,
+    ) -> None:
         await query_loop.start()
         await query_tracker.wait_queries()
-        assert caplog.messages == [
-            'connected to database "db"',
-            'running query "q" on database "db"',
-            'updating metric "m" set 100.0 {database="db"}',
-            re_match(
-                r'updating metric "query_latency" observe .* \{database="db",query="q"\}'
+        assert [
+            log.debug(
+                "updating metric",
+                metric="m",
+                method="set",
+                value=100.0,
+                labels={"database": "db"},
             ),
-            'updating metric "queries" inc 1 {database="db",query="q",status="success"}',
-        ]
+            log.debug(
+                "updating metric",
+                metric="query_latency",
+                method="observe",
+                value=ANY,
+                labels={"database": "db", "query": "q"},
+            ),
+            log.debug(
+                "updating metric",
+                metric="queries",
+                method="inc",
+                value=1,
+                labels={"database": "db", "query": "q", "status": "success"},
+            ),
+        ] <= log.events
 
     async def test_run_query_log_labels(
-        self, caplog, query_tracker, config_data, make_query_loop
-    ):
-        """Debug messages include metric labels."""
+        self,
+        log: StructuredLogCapture,
+        query_tracker: QueryTracker,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+    ) -> None:
         config_data["metrics"]["m"]["labels"] = ["l"]
         config_data["queries"]["q"]["sql"] = 'SELECT 100.0 AS m, "foo" AS l'
         query_loop = make_query_loop()
-        caplog.set_level(logging.DEBUG)
         await query_loop.start()
         await query_tracker.wait_queries()
-        assert caplog.messages == [
-            'connected to database "db"',
-            'running query "q" on database "db"',
-            'updating metric "m" set 100.0 {database="db",l="foo"}',
-            re_match(
-                r'updating metric "query_latency" observe .* \{database="db",query="q"\}'
-            ),
-            'updating metric "queries" inc 1 {database="db",query="q",status="success"}',
-        ]
+        assert log.has(
+            "updating metric",
+            level="debug",
+            metric="m",
+            method="set",
+            value=100.0,
+            labels={"database": "db", "l": "foo"},
+        )
 
     async def test_run_query_increase_database_error_count(
-        self, mocker, query_tracker, config_data, make_query_loop, registry
-    ):
-        """Count of database errors is incremented on failed connection."""
+        self,
+        mocker: MockerFixture,
+        query_tracker: QueryTracker,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+        registry: MetricsRegistry,
+    ) -> None:
+        """Query errors are logged."""
         query_loop = make_query_loop()
         db = query_loop._databases["db"]
-        mock_connect = mocker.patch.object(db._engine, "connect")
+        mock_connect = mocker.patch.object(db._conn.engine, "connect")
         mock_connect.side_effect = Exception("connection failed")
         await query_loop.start()
         await query_tracker.wait_failures()
@@ -337,9 +373,12 @@ class TestQueryLoop:
         assert metric_values(queries_metric) == [1.0]
 
     async def test_run_query_increase_query_error_count(
-        self, query_tracker, config_data, make_query_loop, registry
-    ):
-        """Count of errored queries is incremented on error."""
+        self,
+        query_tracker: QueryTracker,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+        registry: MetricsRegistry,
+    ) -> None:
         config_data["queries"]["q"]["sql"] = "SELECT 100.0 AS a, 200.0 AS b"
         query_loop = make_query_loop()
         await query_loop.start()
@@ -350,19 +389,25 @@ class TestQueryLoop:
         }
 
     async def test_run_query_increase_timeout_count(
-        self, query_tracker, config_data, make_query_loop, registry
-    ):
-        """Count of errored queries is incremented on timeout."""
+        self,
+        mocker: MockerFixture,
+        query_tracker: QueryTracker,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+        registry: MetricsRegistry,
+    ) -> None:
         config_data["queries"]["q"]["timeout"] = 0.1
         query_loop = make_query_loop()
         await query_loop.start()
         db = query_loop._databases["db"]
         await db.connect()
 
-        async def execute(sql, parameters):
+        async def execute(
+            sql: TextClause, parameters: dict[str, t.Any] | None
+        ) -> None:
             await asyncio.sleep(1)  # longer than timeout
 
-        db._conn.execute = execute
+        mocker.patch.object(db._conn, "execute", execute)
 
         await query_tracker.wait_failures()
         queries_metric = registry.get_metric("queries")
@@ -371,9 +416,11 @@ class TestQueryLoop:
         }
 
     async def test_run_query_at_interval(
-        self, advance_time, query_tracker, query_loop
-    ):
-        """Queries are run at the specified time interval."""
+        self,
+        advance_time: AdvanceTime,
+        query_tracker: QueryTracker,
+        query_loop: QueryLoop,
+    ) -> None:
         await query_loop.start()
         await advance_time(0)  # kick the first run
         # the query has been run once
@@ -386,25 +433,33 @@ class TestQueryLoop:
         assert len(query_tracker.queries) == 2
 
     async def test_run_timed_queries_invalid_result_count(
-        self, query_tracker, config_data, make_query_loop, advance_time
-    ):
-        """Timed queries returning invalid elements count are removed."""
+        self,
+        query_tracker: QueryTracker,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+    ) -> None:
         config_data["queries"]["q"]["sql"] = "SELECT 100.0 AS a, 200.0 AS b"
+        config_data["queries"]["q"]["interval"] = 1.0
         query_loop = make_query_loop()
         await query_loop.start()
-        await advance_time(0)  # kick the first run
-        assert len(query_tracker.queries) == 1
+        timed_call = query_loop._timed_calls["q"]
+        await asyncio.sleep(1.1)
+        await query_tracker.wait_failures()
+        assert len(query_tracker.failures) == 1
         assert len(query_tracker.results) == 0
-        # the query is not run again
-        await advance_time(5)
-        assert len(query_tracker.results) == 0
-        await advance_time(5)
+        # the query has been stopped and removed
+        assert not timed_call.running
+        await asyncio.sleep(1.1)
+        await query_tracker.wait_failures()
+        assert len(query_tracker.failures) == 1
         assert len(query_tracker.results) == 0
 
     async def test_run_timed_queries_invalid_result_count_stop_task(
-        self, query_tracker, config_data, make_query_loop
-    ):
-        """Timed queries returning invalid result counts are stopped."""
+        self,
+        query_tracker: QueryTracker,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+    ) -> None:
         config_data["queries"]["q"]["sql"] = "SELECT 100.0 AS a, 200.0 AS b"
         config_data["queries"]["q"]["interval"] = 1.0
         query_loop = make_query_loop()
@@ -417,9 +472,12 @@ class TestQueryLoop:
         assert query_loop._timed_calls == {}
 
     async def test_run_timed_queries_not_removed_if_not_failing_on_all_dbs(
-        self, tmp_path, query_tracker, config_data, make_query_loop
-    ):
-        """Timed queries are removed when they fail on all databases."""
+        self,
+        tmp_path: Path,
+        query_tracker: QueryTracker,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+    ) -> None:
         db1 = tmp_path / "db1.sqlite"
         db2 = tmp_path / "db2.sqlite"
         config_data["databases"] = {
@@ -457,9 +515,11 @@ class TestQueryLoop:
         assert len(query_tracker.failures) == 1
 
     async def test_run_aperiodic_queries(
-        self, query_tracker, config_data, make_query_loop
-    ):
-        """Queries with null interval can be run explicitly."""
+        self,
+        query_tracker: QueryTracker,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+    ) -> None:
         del config_data["queries"]["q"]["interval"]
         query_loop = make_query_loop()
         await query_loop.run_aperiodic_queries()
@@ -468,9 +528,11 @@ class TestQueryLoop:
         assert len(query_tracker.queries) == 2
 
     async def test_run_aperiodic_queries_invalid_result_count(
-        self, query_tracker, config_data, make_query_loop
-    ):
-        """Aperiodic queries returning invalid elements count are removed."""
+        self,
+        query_tracker: QueryTracker,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+    ) -> None:
         config_data["queries"]["q"]["sql"] = "SELECT 100.0 AS a, 200.0 AS b"
         del config_data["queries"]["q"]["interval"]
         query_loop = make_query_loop()
@@ -482,9 +544,12 @@ class TestQueryLoop:
         assert len(query_tracker.queries) == 1
 
     async def test_run_aperiodic_queries_not_removed_if_not_failing_on_all_dbs(
-        self, tmp_path, query_tracker, config_data, make_query_loop
-    ):
-        """Periodic queries are removed when they fail on all databases."""
+        self,
+        tmp_path: Path,
+        query_tracker: QueryTracker,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+    ) -> None:
         db1 = tmp_path / "db1.sqlite"
         db2 = tmp_path / "db2.sqlite"
         config_data["databases"] = {
@@ -522,14 +587,13 @@ class TestQueryLoop:
 
     async def test_clear_expired_series(
         self,
-        mocker,
-        tmp_path,
-        query_tracker,
-        config_data,
-        make_query_loop,
-        registry,
-    ):
-        """The query loop clears out series last seen earlier than the specified interval."""
+        tmp_path: Path,
+        advance_time: AdvanceTime,
+        query_tracker: QueryTracker,
+        config_data: dict[str, t.Any],
+        make_query_loop: MakeQueryLoop,
+        registry: MetricsRegistry,
+    ) -> None:
         db = tmp_path / "db.sqlite"
         config_data["databases"]["db"]["dsn"] = f"sqlite:///{db}"
         config_data["metrics"]["m"].update(
@@ -538,7 +602,6 @@ class TestQueryLoop:
                 "expiration": 10,
             }
         )
-        # call metric collection directly
         config_data["queries"]["q"]["sql"] = "SELECT * FROM test"
         del config_data["queries"]["q"]["interval"]
 
@@ -549,24 +612,19 @@ class TestQueryLoop:
             'INSERT INTO test VALUES (20, "bar")',
         )
         query_loop = make_query_loop()
-        mock_timestamp = mocker.patch.object(query_loop, "_timestamp")
-        mock_timestamp.return_value = datetime.now().timestamp()
         await query_loop.run_aperiodic_queries()
         await query_tracker.wait_results()
         queries_metric = registry.get_metric("m")
-        assert metric_values(queries_metric, by_labels=("l",)), {
+        assert metric_values(queries_metric, by_labels=("l",)) == {
             ("foo",): 10.0,
             ("bar",): 20.0,
         }
-        await run_queries(
-            db,
-            "DELETE FROM test WHERE m = 10",
-        )
-        # mock that more time has passed than expiration
-        mock_timestamp.return_value += 20
+        await run_queries(db, "DELETE FROM test WHERE m = 10")
+        # go beyond expiration time
+        await advance_time(20)
         await query_loop.run_aperiodic_queries()
         await query_tracker.wait_results()
         query_loop.clear_expired_series()
-        assert metric_values(queries_metric, by_labels=("l",)), {
+        assert metric_values(queries_metric, by_labels=("l",)) == {
             ("bar",): 20.0,
         }
